% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_br_snakes_atlas.r
\docType{data}
\name{br_snakes_atlas}
\alias{br_snakes_atlas}
\title{Example dataset: br_snakes_atlas}
\format{
A dataframe with 411 rows and 3 colums:
\describe{
\item{species}{A character vector with known current valid name for all reptile species cataloged in The Reptile Database website}
\item{latitude}{Latitude coordinates in decimal degrees}
\item{longitude}{Longitude coordinates in decimal degrees}
}
}
\source{
Nogueira et al. (2019) supplementary material, Table S3
}
\usage{
br_snakes_atlas
}
\description{
This dataset contains Table S3, the georeferenced type localities, from the supplementary material provided by Nogueira et al., (2019)
}
\references{
Nogueira, C. C., Argôlo, A. J. S., Arzamendia, V., Azevedo, J. A. R., Barbo, F. E., Bérnils, R. S., … & Martins, M. (2019). Atlas of Brazilian Snakes: Verified Point-Locality Maps to Mitigate the Wallacean Shortfall in a Megadiverse Snake Fauna. South American Journal of Herpetology, 14(sp1), 1–274. http://dx.doi.org/10.2994/sajh-d-19-00120.1
}
\keyword{datasets}
