% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpecies.R
\name{getSpecies}
\alias{getSpecies}
\title{Building url species list}
\usage{
getSpecies(url, higherTaxa = FALSE)
}
\arguments{
\item{url}{a \emph{character} string with the url from an advanced search in Reptile Database}

\item{higherTaxa}{a \emph{logical} value indicating if user wants the families for each species within the resulting data frame}
}
\value{
if \emph{higherTaxa} = FALSE (default), '\emph{getSpecies}' returns a dataframe with columns: species and url.

if \emph{higherTaxa} = TRUE, '\emph{getSpecies}' returns a dataframe with columns: family, genus, species and url
}
\description{
creates a \emph{dataframe} containing list of species based on a Reptile Database advanced search with their respective url
}
\references{
Liedtke, H. C. (2018). AmphiNom: an amphibian systematics tool. \emph{Systematics and Biodiversity}, 17(1), 1–6. https://doi.org/10.1080/14772000.2018.1518935
}
