% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herpAdvancedSearch.R
\name{herpAdvancedSearch}
\alias{herpAdvancedSearch}
\title{Search The Reptile Database website (TRD): Advanced}
\usage{
herpAdvancedSearch(higher=NULL,
                          genus=NULL,
                          year=NULL,
                          common_name=NULL,
                          synonym=NULL,
                          location=NULL)
}
\arguments{
\item{higher}{A character string with the current valid name of a given reptile higher taxa above genus (e.g.: "snake" or "Boidae")}

\item{genus}{A character string with the current valid name of a given reptile genus (e.g.: "\emph{Apostolepis}")}

\item{year}{A character string to be used as a filter for the year of description of the searched species (e.g.: "2025")}

\item{common_name}{A character string with name potentially regarded as a synonym of a given reptile genus (e.g.: "\emph{Boa diviniloqua}")}

\item{synonym}{A character string with name potentially regarded as a synonym of a given reptile genus (e.g.: "\emph{Boa diviniloqua}")}

\item{location}{A character string with a location from which the user wants the list of species expected to occur}
}
\value{
the url to be used in herpSpecies()
}
\description{
Creates the link for multiple species sampling by herpSpecies()
}
\examples{
herpAdvancedSearch(higher = "snakes", year = "2010", location = "Brazil")
herpAdvancedSearch(year = "2010 OR 2011 OR 2012")
herpAdvancedSearch(genus = "Apostolepis OR \"Boa\" OR Atractus") #to quote one

}
