% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herpSplitCheck.R
\name{herpSplitCheck}
\alias{herpSplitCheck}
\title{Check for possible splits in herp species names}
\usage{
herpSplitCheck(
  x,
  pubDate = NULL,
  verbose = TRUE,
  cores = max(1, parallel::detectCores() - 1),
  showProgress = TRUE
)
}
\arguments{
\item{x}{A character vector of species names to check.}

\item{pubDate}{An integer year (e.g., 2019) used to detect new taxa.}

\item{verbose}{Logical; whether to print progress.}

\item{cores}{Integer. Number of CPU cores to use for parallel processing. Default is one less than the total available cores.}

\item{showProgress}{Logical. If \code{TRUE}, prints data sampling progress. Default is \code{TRUE}.}
}
\value{
A data frame with columns: query, species, status.
}
\description{
Check for possible splits in herp species names
}
