% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herpSync.R
\name{herpSync}
\alias{herpSync}
\title{Synchronize species names using The Reptile Database}
\usage{
herpSync(x, solveAmbiguity = TRUE, cores = max(1, parallel::detectCores() - 1), verbose = TRUE, showProgress = TRUE, getLink = FALSE)
}
\arguments{
\item{x}{A character vector of taxon names to be processed (e.g., species list, phylogenetic tip labels, or trait table entries).}

\item{solveAmbiguity}{Logical. If \code{TRUE}, samples the synonyms of species with ambiguous nomenclature}

\item{cores}{Integer. Number of CPU cores to use for parallel processing. Default is one less than the number of available cores.}

\item{verbose}{Logical. If \code{TRUE}, prints species information in the console. Default is \code{TRUE}.}

\item{showProgress}{Logical. If \code{TRUE}, prints sampling progress in the console. Default is \code{FALSE}.}

\item{getLink}{Logical. If \code{TRUE}, includes the RDB URL for each valid species names detected Default is \code{FALSE}.}
}
\value{
A data frame with the following columns:
\itemize{
\item \code{query}: original input names from the query.
\item \code{RDB}: best-matching valid names according to the RDB.
\item \code{status}: description of the outcome (e.g., \code{up_to_date}, \code{"updated"}, \code{"ambiguous"},  \code{"not_found"}).
\item \code{url}: the url searched for each querried species. The urls of species with ambiguous nomenclature returns a list of species of which the querried species is considered a synonym
}
}
\description{
Queries a list of species parsed to \code{herpSearch} and returns a data frame with current valid names for queried species
}
\note{
\code{herpSync} does not make taxonomic claims.
The function compares queried species with valid species names and returns "up_to_date" for any binomial that is considered valid,
regardless of if it has been split and the name applies only to a specific population.
}
\examples{
query <- c("Vieira-Alencar authoristicus", "Boa atlantica", "Boa diviniloqua", "Boa imperator")
\donttest{
herpSync(x=query)
}


}
\references{
Liedtke, H. C. (2018). AmphiNom: an amphibian systematics tool. \emph{Systematics and Biodiversity}, 17(1), 1–6. https://doi.org/10.1080/14772000.2018.1518935
}
