% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herpTidySyn.R
\name{herpTidySyn}
\alias{herpTidySyn}
\title{Printing herpSync outcome in a tidy way}
\usage{
herpTidySyn(df, filter = NULL)
}
\arguments{
\item{df}{The data frame derivated from \code{herpSync}}

\item{filter}{Logical. If \code{TRUE} will print only the species entries with unresolved nomenclature (e.g.: ambiguous or unknown). Default is \code{TRUE}}
}
\value{
Invisibly returns \code{NULL}. Used for side-effect printing only.
}
\description{
Prints the data frame derivated from \code{herpSync} in a tidy way.
Optionally, it filters the data frame for species with unresolved nomenclature only.
}
\examples{
df <- data.frame(
  species = c("Genus epithet 1", 
              "Genus epithet 2",
              "Genus epithet 3",
              "Genus epithet 4",
              "Genus epithet 5"),
  synonyms = c("Genus epithet 1.1; Genus epithet 1.2",
                "Genus epithet 2",
                "Genus epithet 3",
                "Not found",
                "Genus epithet 5.1; Genus epithet 5.2; Genus epithet 5.3"),
  status = c("ambiguous", 
             "updated",
             "up_to_date",
             "unknown",
             "ambiguous"),
  stringsAsFactors = FALSE
)
herpTidySyn(df, filter = TRUE)

}
