% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letsHerp-package.R
\docType{package}
\name{letsHerp-package}
\alias{letsHerp}
\alias{letsHerp-package}
\title{letsHerp: What the Package Does (One Line, Title Case)}
\description{
The package compiles reptile species url from The Reptile Dabase website and summarizes species synonyms in a dataframe.
}
\author{
\strong{Maintainer}: Joao Paulo dos Santos Vieira-Alencar \email{joaopaulo.valencar@gmail.com}

}
\keyword{internal}
