% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letsHerp-package.R
\docType{package}
\name{letsHerp}
\alias{letsHerp-package}
\alias{letsHerp}
\title{letsHerp: An R Interface to the Reptile Database}
\description{
The letsHerp package provides functions to interact with and extract taxonomic and distributional data
from the Reptile Database (https://reptile-database.reptarium.cz). It supports advanced search queries,
extraction of species names, synonyms, higher taxa, and related metadata for herpetological research.
}
\author{
\strong{Maintainer}: Joao Paulo dos Santos Vieira-Alencar \email{joaopaulo.valencar@gmail.com}

}
\keyword{internal}
