% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letsRept-package.R
\docType{package}
\name{letsRept}
\alias{letsRept-package}
\alias{letsRept}
\title{letsRept: An Interface to the Reptile Database}
\description{
The letsRept package provides functions to interact with and extract taxonomic and distributional data
from the Reptile Database (https://reptile-database.reptarium.cz). It supports advanced search queries,
extraction of species names, synonyms, higher taxa, and related metadata for herpetological research.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/joao-svalencar/letsRept}
  \item Report bugs at \url{https://github.com/joao-svalencar/letsRept/issues}
}

}
\author{
\strong{Maintainer}: João Paulo dos Santos Vieira-Alencar \email{joaopaulo.valencar@gmail.com} (\href{https://orcid.org/0000-0001-6894-6773}{ORCID})

Authors:
\itemize{
  \item Christoph Liedtke \email{christoph.liedtke@ebd.csic.es} (\href{https://orcid.org/0000-0002-6221-8043}{ORCID})
}

}
\keyword{internal}
