% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{match_taxon}
\alias{match_taxon}
\title{Matching higher taxa information}
\usage{
match_taxon(taxa_vector, rank_list)
}
\arguments{
\item{taxa_vector}{A character vector containing taxonomic information.}

\item{rank_list}{A character vector of taxonomic rank names to match.}
}
\value{
A character string of matched ranks, comma-separated. Returns NA if none found.
}
\description{
A helper function to herpSpecies(). Identify species orders or suborders from taxa_vector
}
\keyword{internal}
