% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reptRefs.R
\name{reptRefs}
\alias{reptRefs}
\title{Extract references from the Reptile Database}
\usage{
reptRefs(x = NULL, getLink = TRUE)
}
\arguments{
\item{x}{A binomial species name (e.g., "Boa constrictor").}

\item{getLink}{Logical; if TRUE, also returns associated links to references.}
}
\value{
A character vector of references (if getLink = FALSE), or a data frame with columns \code{reference} and \code{link}.
}
\description{
Extract references from the Reptile Database
}
\examples{
\donttest{
df <- reptRefs("Apostolepis adhara")
}

}
