% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{safeParallel}
\alias{safeParallel}
\title{Applying parallel sampling}
\usage{
safeParallel(data, FUN, cores = cores)
}
\arguments{
\item{data}{species_list to be sampled}

\item{FUN}{function to be sampled in parallel}

\item{cores}{number of cores passed to main function}
}
\value{
the result of FUN performed in parallel running within user defined number of cores
}
\description{
A helper function to herpSpecies(). Decides which parallel template to follow according to user OS
}
\keyword{internal}
