% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchHerp.R
\name{searchHerp}
\alias{searchHerp}
\title{Search The Reptile Database website (TRD)}
\usage{
searchHerp(
  binomial = NULL,
  highertaxa = NULL,
  genus = NULL,
  distribution = NULL
)
}
\arguments{
\item{binomial}{a \emph{character} string with the current valid binomial name of a given reptile species (e.g.: "Apostolepis adhara")}

\item{highertaxa}{a \emph{character} string with the current valid name of a given reptile higher taxa above genus (e.g.: "snake" or "Boidae")}

\item{genus}{a \emph{character} string with the current valid name of a given reptile genus (e.g.: "Apostolepis")}

\item{distribution}{a \emph{character} string with a location from which the user wants the list of species expected to occur}
}
\value{
If a search for an specific species information (e.g.: argument binomial is not NULL): returns the species information collected from its respective page in The Reptile Database
If advanced search: returns the url to be used in getSpecies()
}
\description{
ATTENTION: under development, may not work yet. Alternatively, for advanced search copy the link with the results from TRD and run getSpecies()
simulates searches in The Reptile Database website and provides the data of a single species or the url for multiple species sampling by getSpecies()
}
