% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allReptiles}
\alias{allReptiles}
\title{Example dataset: allReptiles}
\format{
A dataframe with 12384 rows and 2 variables:
\describe{
\item{species}{A character vector with known current valid name for all reptile species cataloged in The Reptile Database website (download: May 15th, 2025)}
\item{url}{A character column with the respective url to access all reptile species cataloged in The Reptile Database website information (download: May 15th, 2025)}
}
}
\source{
The data was sampled from The Reptile Database website https://reptile-database.reptarium.cz using function HerpNom::getSpecies() with the url obtained from an 'Advanced search' set to exclude all reptile species described to the fictional planet Arrakis (-Arrakis).
}
\usage{
allReptiles
}
\description{
This dataset contains some toy data.
}
\keyword{datasets}
