% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_synonymsRef.R
\docType{data}
\name{allSynonymsRef}
\alias{allSynonymsRef}
\title{Example dataset: allSynonymsRef}
\format{
A dataframe with 111,549 rows and 3 variables:
\describe{
\item{species}{A character vector with known current valid name for all reptile species cataloged in The Reptile Database website (download: September 18tb, 2025)}
\item{synonym}{A character column with the respective synonyms for all reptile species cataloged in The Reptile Database website information (download: September 18tb, 2025)}
\item{ref}{A charater column with the synonyms and respective references that used it}
}
}
\source{
The data was sampled from The Reptile Database website https://reptile-database.reptarium.cz using function letsRept::reptSynonyms(letsRept::allReptiles)
}
\usage{
allSynonymsRef
}
\description{
This dataset contains the valid names and respective listed synonyms for all reptile species cataloged in The Reptile Database.
}
\keyword{datasets}
