% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herpAdvancedSearch.R
\name{herpAdvancedSearch}
\alias{herpAdvancedSearch}
\title{Search The Reptile Database website (TRD): Advanced}
\usage{
herpAdvancedSearch(highertaxa=NULL, genus=NULL, synonym=NULL, distribution=NULL)
}
\arguments{
\item{highertaxa}{A character string with the current valid name of a given reptile higher taxa above genus (e.g.: "snake" or "Boidae")}

\item{genus}{A character string with the current valid name of a given reptile genus (e.g.: "\emph{Apostolepis}")}

\item{synonym}{A character string with name potentially regarded as a synonym of a given reptile genus (e.g.: "\emph{Boa diviniloqua}")}

\item{distribution}{A character string with a location from which the user wants the list of species expected to occur}
}
\value{
the url to be used in herpSpecies()
}
\description{
Searches The Reptile Database website and provides the data of a single species or the url for multiple species sampling by herpSpecies()\\cr
#' \emph{ATTENTION:} under development, may not work yet.\\cr
Alternatively, for advanced search copy the link with the results from TRD and run herpSpecies()
}
