% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herpSearch.R
\name{herpSearch}
\alias{herpSearch}
\title{Search The Reptile Database website (TRD): Simple}
\usage{
herpSearch(binomial=NULL)
}
\arguments{
\item{binomial}{a \emph{character} string with the current valid binomial name of a given reptile species (e.g.: "\emph{Apostolepis adhara}")}
}
\value{
Returns species information collected from its respective page in The Reptile Database.
}
\description{
Simulates searches in The Reptile Database website and provides the data of a single species or the url for multiple species sampling by herpSpecies()\\cr
#' \emph{ATTENTION:} under development, may not work yet.\\cr
Alternatively, for advanced search copy the link with the results from TRD and run herpSpecies()
}
