% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herpSearch.R
\name{herpSearch}
\alias{herpSearch}
\title{Search The Reptile Database website (TRD): Simple}
\usage{
herpSearch(binomial=NULL)
}
\arguments{
\item{binomial}{A character string with the current valid binomial name of a given reptile species.}
}
\value{
Returns species information as provided by The Reptile Database.
}
\description{
Searches The Reptile Database website and provides the data of a single species
}
\examples{
herpSearch("Boa constrictor")

}
