% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herpSpecies.R
\name{herpSpecies}
\alias{herpSpecies}
\title{Reptile species summary}
\usage{
herpSpecies(url,
                   dataList = NULL, 
                   taxonomicInfo=FALSE, 
                   fullHigher=FALSE, 
                   getLink=FALSE,
                   cores = max(1, parallel::detectCores() - 1),
                   checkpoint = NULL,
                   backup_file = NULL
                   )
}
\arguments{
\item{url}{A character string with the url from an advanced search in Reptile Database website or from letsHerp::herpAdvancedSearch.}

\item{dataList}{A data frame with columns: species and url, only for sampling taxonomicInfo from already sampled species links.}

\item{taxonomicInfo}{A logical value indicating if user wants full species taxonomic information (specifically: Order, Suborder, Family, Genus, species author and description year) for each species. default = \emph{TRUE}}

\item{fullHigher}{A logical value indicating if user wants the full higher taxa information (including e.g.: subfamily) for each species, as available in The Reptile Database website (e.g. single character string). default = \emph{FALSE}. OBS.: Requires taxonomicInfo = TRUE}

\item{getLink}{A logical value indicating if user wants the url that provides access to each species information (e.g: to use with herpSynonyms()). default = \emph{TRUE}}

\item{cores}{An integer representing how many cores to use during parallel sampling. default is 1 less than all available cores}

\item{checkpoint}{An integer representing the number of species to process before saving progress to the backup file.
Helps prevent data loss in case the function stops unexpectedly. Backups are saved only if cores = 1 and if checkpoint is not NULL.
OBS.: If set to 1, progress will be saved after every species (safe but slower).}

\item{backup_file}{A character string with the Path to an \code{.rds} file where intermediate results will be saved if \code{checkpoint} is not NULL. Must end in \code{.rds}.}
}
\value{
if taxonomicInfo = FALSE (default), the function returns a vector with the list of species

if taxonomicInfo = TRUE, the function returns a data frame with columns: order, suborder (e.g.: Sauria or Serpentes only; when available), family, genus, species, author and year

Optionally, the function might return a data frame with a column with the full higher taxa information as reported in The Reptile Database, and the species respective url (necessary if looking for synonyms afterwards)
}
\description{
Creates a data frame containing higher taxa information for a list of reptile species based on a Reptile Database advanced search optionally with their respective url.
}
