% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herpSpecies.R
\name{herpSpecies}
\alias{herpSpecies}
\title{Building url species list}
\usage{
herpSpecies(url, higherTaxa = TRUE, fullHigher = FALSE, getlink = FALSE)
}
\arguments{
\item{url}{A \emph{character} string with the url from an advanced search in Reptile Database.}

\item{higherTaxa}{A \emph{logical} value indicating if user wants higher taxa information (specifically: Order, Suborder, Family and Genus) for each species retrieved. default = \emph{TRUE}.}

\item{fullHigher}{A \emph{logical} value indicating if user wants the full higher taxa information (including e.g.: subfamily) for each species retrieved, as available in The Reptile Database website (e.g. single character string). default = \emph{FALSE}.}

\item{getlink}{A \emph{logical} value indicating if user wants the url that provides access to each retrieved species (e.g: to use with herpSynonyms()). default = \emph{TRUE}}
}
\value{
if \emph{higherTaxa} = FALSE (default), '\emph{herpSpecies}' returns a dataframe with columns: species and url.

if \emph{higherTaxa} = TRUE, '\emph{herpSpecies}' returns a dataframe with columns: family, genus, species and url
}
\description{
Creates a \emph{dataframe} containing a list of species based on a Reptile Database advanced search with their respective url
}
\references{
Liedtke, H. C. (2018). AmphiNom: an amphibian systematics tool. \emph{Systematics and Biodiversity}, 17(1), 1–6. https://doi.org/10.1080/14772000.2018.1518935
}
