% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herpSynonyms.R
\name{herpSynonyms}
\alias{herpSynonyms}
\title{Get reptile species synonyms}
\usage{
herpSynonyms(x,
                    getRef = FALSE,
                    checkpoint = NULL,
                    backup_file = NULL,
                    resume=FALSE,
                    cores = max(1, parallel::detectCores() - 1))
}
\arguments{
\item{x}{A data frame with columns: 'species' and 'url' (their respective Reptile Database url).
Could be the output of letsHerp::herpSpecies().}

\item{getRef}{A logical value. If TRUE, returns synonyms with the respective references that mention them. default = \emph{FALSE}}

\item{checkpoint}{An integer representing the number of species to process before saving progress to the backup file.
Helps prevent data loss in case the function stops unexpectedly. Backups are saved only if checkpoint is not NULL.
OBS.: checkpoint are not used under parallel processing, if mandatory choose cores = 1 (safe, but decreases processing performance).}

\item{backup_file}{A character string with the path to access, read and save the backup file.}

\item{resume}{A logical value. If TRUE, takes the path to backup_file and resume sampling from the last backup file.
OBS.: backup_file are not created under parallel processing, if mandatory choose cores = 1 (safe, but decreases processing performance).}

\item{cores}{number of cores dedicated to parallel processing}
}
\value{
'herpSynonyms' returns a data frame with columns: species and their respective synonyms according to the version of Reptile Database. Optionally, returns the references that mentioned each synonym.
}
\description{
creates a data frame containing a list of reptile species current valid names according to The Reptile Database alongside with all their recognized synonyms
}
\examples{
boa <- letsHerp::allReptiles[grep("^Boa\\\\s", allReptiles$species),]
boa_syn <- herpSynonyms(boa, getRef = FALSE, cores=2) #only synonyms

}
\references{
Uetz, P., Freed, P, Aguilar, R., Reyes, F., Kudera, J. & Hošek, J. (eds.) (2025) The Reptile Database, http://www.reptile-database.org
Liedtke, H. C. (2018). AmphiNom: an amphibian systematics tool. \emph{Systematics and Biodiversity}, 17(1), 1–6. https://doi.org/10.1080/14772000.2018.1518935
}
