% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herpSynonyms.R
\name{herpSynonyms}
\alias{herpSynonyms}
\title{Get reptile species synonyms}
\usage{
herpSynonyms(x, 
                    batch_size = NULL,
                    resume=FALSE,
                    backup_file = NULL,
                    getRef = FALSE)
}
\arguments{
\item{x}{A data frame with columns: 'species' and 'url' (their respective Reptile Database url).
Could be the output of letsHerp::herpSpecies().}

\item{batch_size}{An integer representing the number of species to process before saving progress to the backup file.
Helps prevent data loss in case the function stops unexpectedly. Backups are saved only if batch_size is not NULL.}

\item{resume}{A logical value. If TRUE, takes the path to backup_file and resume sampling from the last backup file.}

\item{backup_file}{A character string with the path to access, read and save the backup file.}

\item{getRef}{A logical value. If TRUE, returns synonyms with the respective references that mention them. default = \emph{FALSE}}
}
\value{
'herpSynonyms' returns a data frame with columns: species and their respective synonyms according to the version of Reptile Database. Optionally, returns the references that mentioned each synonym.
}
\description{
creates a data frame containing a list of reptile species current valid names according to The Reptile Database alongside with all their recognized synonyms
}
\examples{
boaLink <- herpAdvancedSearch(genus = "Boa") #creates advanced search link
boa <- herpSpecies(boaLink, getLink=TRUE, taxonomicInfo = FALSE)
boa_syn <- herpSynonyms(boa, getRef = TRUE) #synonyms with respective references
boa_syn <- herpSynonyms(boa, getRef = FALSE) #only synonyms

}
\references{
Uetz, P., Freed, P, Aguilar, R., Reyes, F., Kudera, J. & Hošek, J. (eds.) (2025) The Reptile Database, http://www.reptile-database.org
Liedtke, H. C. (2018). AmphiNom: an amphibian systematics tool. \emph{Systematics and Biodiversity}, 17(1), 1–6. https://doi.org/10.1080/14772000.2018.1518935
}
