% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herpSynonyms.R
\name{herpSynonyms}
\alias{herpSynonyms}
\title{Get reptile species synonyms}
\usage{
herpSynonyms(x)
}
\arguments{
\item{x}{a \emph{dataframe} columns: 'species' and 'url' (their respective Reptile Database url). Could be the output of letsHerp::herpSpecies()}
}
\value{
'\emph{herpSynonyms}' returns a dataframe with columns: species and their respective synonyms according to the version of Reptile Database.
}
\description{
creates a \emph{dataframe} containing a list of reptile species current valid names according to The Reptile Database alongside with all their recognized synonyms
}
\references{
Uetz, P., Freed, P, Aguilar, R., Reyes, F., Kudera, J. & Hošek, J. (eds.) (2025) The Reptile Database, http://www.reptile-database.org
Liedtke, H. C. (2018). AmphiNom: an amphibian systematics tool. \emph{Systematics and Biodiversity}, 17(1), 1–6. https://doi.org/10.1080/14772000.2018.1518935
}
